// frontend/src/types/user.ts
export type UserRole = "CUSTOMER" | "ARTISAN" | "ADMIN";
export type UserType = "CUSTOMER" | "ARTISAN" | "ADMIN";
export type AdminPermissionLevel = "SUPER_ADMIN" | "MANAGER" | "SUPPORT";

export type UserResponseDTO = {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  storeId: string | null;
  isDeleted: boolean;
  createdAt: string;
  updatedAt: string;
};